Occlusion culling works!
Removed second pass as it doesn't seem to be needed. Sampling the maximum depth
instead of the minimum for the depth pyramid. Sampling a 3x3 neighboring
pixels instead of 2x2 fixed the flickering. In the Menger's sponge
scene framerate is double, as expected, since more than half the geometry
is usually occluded.