Add SVO class (untested) - Construct GPU buffers for nodes and voxels
I need to incorporate the nodes and voxels created by the SVO into the
vertex buffer. To make it work with removing and adding voxels I will
need to edit the occlusion culling shaders to create the draw commands
using the SVO. Hence the childrenStartIndex and voxelStartIndex in the
GPUNode struct.