Update the SVO
I will not flatten the buffers. I will try to draw using the SVO, by
creating the commands traversing the tree in the culling shader. I am
thinking of a design where the tree will be edited in the GPU, when
I implement editing. I still have functions on the CPU to remove or add
voxels that I will probably use in the future if I want to add
multi-player but I'm not sure if they're working. What a headache...