Add Materials class - Create buffer for all light sources
Save each light in a linear buffer split in chunks. Have some add and
remove light function for the future if I edit the terrain. I don't know
how bad it is to reconstruct the whole linear buffer. If I add or remove
a light I am shifting all the linear buffer.